myproject/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
gunicorn
whitenoise
celery[redis]==5.3.6
pandas
openpyxl 
